+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- [#101](https://github.com/diprokon/ng-table-virtual-scroll/issues/101):
+  Sticky footer
+
 ## [1.5.0] - 2023-02-05
 
 ### Added